move getConcreteEntiyWhere as util
