+++ packages/graphql/src/translate/queryAST/factory/OperationFactory.ts
@@ -40,6 +40,7 @@ import type { InterfaceEntityAdapter } from "../../../schema-model/entity/model-
 import { InterfaceReadOperation } from "../ast/operations/interfaces/InterfaceReadOperation";
 import { InterfaceReadPartial } from "../ast/operations/interfaces/InterfaceReadPartial";
 import { isUnionEntity } from "../utils/is-union-entity";
+import { getConcreteWhere } from "../utils/get-concrete-where";
 
 export class OperationsFactory {
     private filterFactory: FilterFactory;
@@ -88,7 +89,7 @@ export class OperationsFactory {
                     target: concreteEntity,
                 });
 
-                const whereArgs = this.getConcreteWhere(resolveTreeWhere, entity, concreteEntity);
+                const whereArgs = getConcreteWhere(resolveTreeWhere, entity, concreteEntity);
 
                 return this.hydrateReadOperation({
                     entity: concreteEntity,
@@ -124,35 +125,6 @@ export class OperationsFactory {
                 operation.addPagination(pagination);
             }
         }
-    }
-
-    /** 
-     * Given a Record<string, any> representing a where argument for a composite target fields returns its concrete where argument.
-        For instance, given:
-        {
-            Genre: { name: "Horror" },
-            Movie: { title: "The Matrix" }
-        } 
-        Returns { name: "Horror" } if the concreteTarget is Genre.
-     **/
-    private getConcreteWhere(
-        whereArgs: Record<string, any>,
-        compositeTarget: UnionEntityAdapter | InterfaceEntityAdapter,
-        concreteTarget: ConcreteEntityAdapter
-    ): Record<string, any> {
-        if (whereArgs) {
-            if (isUnionEntity(compositeTarget)) {
-                return whereArgs[concreteTarget.name] ?? {};
-            } else {
-                // interface may have shared filters, inject them as if they were present under _on
-                const sharedInterfaceFilters = Object.entries(whereArgs).filter(([key]) => key !== "_on");
-
-                return whereArgs["_on"]
-                    ? { ...Object.fromEntries(sharedInterfaceFilters), ...whereArgs["_on"][concreteTarget.name] }
-                    : Object.fromEntries(sharedInterfaceFilters);
-            }
-        }
-        return {};
     }
 
     // TODO: dupe from read operation
